requirements.txt
Flask==2.0.1
Cerberus==1.3.4
Flask_JWT_Extended==4.2.3
Flask_Cors==3.0.10
requests==2.26.0
flask_lambda==0.0.4
asposestorage==1.0.2
python-dotenv==0.19.0
